refactor the restart kubelet to extract PerNodePod commands
