+++ pkg/cli/clusteroperator/pernodepod/per-node-namespace.yaml
@@ -1,7 +1,6 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  generateName: openshift-restart-kubelet-
   labels:
     pod-security.kubernetes.io/audit: privileged
     pod-security.kubernetes.io/enforce: privileged
