+++ pkg/cli/clusteroperator/restartkubelet/command.go
@@ -2,20 +2,16 @@ package restartkubelet
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"os/signal"
-	"strconv"
-	"strings"
 	"syscall"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	corev1 "k8s.io/api/core/v1"
 
-	imagev1client "github.com/openshift/client-go/image/clientset/versioned/typed/image/v1"
-	"github.com/openshift/library-go/pkg/image/imageutil"
+	"github.com/openshift/oc/pkg/cli/clusteroperator/pernodepod"
+
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	"k8s.io/client-go/kubernetes"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 	"k8s.io/kubectl/pkg/util/i18n"
 	"k8s.io/kubectl/pkg/util/templates"
@@ -46,30 +42,25 @@ var (
 )
 
 type RestartKubeletOptions struct {
-	RESTClientGetter     genericclioptions.RESTClientGetter
-	PrintFlags           *genericclioptions.PrintFlags
-	ResourceBuilderFlags *genericclioptions.ResourceBuilderFlags
-
-	// TODO push this into genericclioptions
-	DryRun bool
-
-	Parallelism string
+	PerNodePodOptions *pernodepod.PerNodePodOptions
 
 	genericclioptions.IOStreams
 }
 
 func NewRestartKubelet(restClientGetter genericclioptions.RESTClientGetter, streams genericclioptions.IOStreams) *RestartKubeletOptions {
 	return &RestartKubeletOptions{
-		RESTClientGetter: restClientGetter,
-		PrintFlags:       genericclioptions.NewPrintFlags("regeneration set"),
-		ResourceBuilderFlags: genericclioptions.NewResourceBuilderFlags().
-			WithLabelSelector("").
-			WithFieldSelector("").
-			WithAll(false).
-			WithLocal(false).
-			WithLatest(),
-
-		Parallelism: "10%",
+		PerNodePodOptions: pernodepod.NewPerNodePodOptions(
+			"openshift-restart-kubelet-",
+			func(ctx context.Context, namespaceName, nodeName, imagePullSpec string) (*corev1.Pod, error) {
+				restartObj := pod.DeepCopy()
+				restartObj.Namespace = namespaceName
+				restartObj.Spec.NodeName = nodeName
+				restartObj.Spec.Containers[0].Image = imagePullSpec
+				return restartObj, nil
+			},
+			restClientGetter,
+			streams,
+		),
 
 		IOStreams: streams,
 	}
@@ -110,74 +101,15 @@ func NewCmdRestartKubelet(restClientGetter genericclioptions.RESTClientGetter, s
 
 // AddFlags registers flags for a cli
 func (o *RestartKubeletOptions) AddFlags(cmd *cobra.Command) {
-	o.PrintFlags.AddFlags(cmd)
-	o.ResourceBuilderFlags.AddFlags(cmd.Flags())
-
-	cmd.Flags().BoolVar(&o.DryRun, "dry-run", o.DryRun, "Set to true to use server-side dry run.")
-	cmd.Flags().StringVar(&o.Parallelism, "parallelism", o.Parallelism, "parallelism is a raw number or a percentage of the nodes we need to restart at the same time.")
+	o.PerNodePodOptions.AddFlags(cmd)
 }
 
 func (o *RestartKubeletOptions) ToRuntime(args []string) (*RestartKubeletRuntime, error) {
-	parallelPercentage := int64(0)
-	parallelInt, intParseErr := strconv.ParseInt(o.Parallelism, 10, 32)
-	if intParseErr != nil {
-		if !strings.HasSuffix(o.Parallelism, "%") {
-			return nil, fmt.Errorf("--parallelism must be either N or N%%: %w", intParseErr)
-		}
-		percentageString := o.Parallelism[0 : len(o.Parallelism)-1]
-		var percentParseErr error
-		parallelPercentage, percentParseErr = strconv.ParseInt(percentageString, 10, 32)
-		if percentParseErr != nil {
-			return nil, fmt.Errorf("--parallelism must be either N or N%%: %w", intParseErr)
-		}
-	}
-
-	printer, err := o.PrintFlags.ToPrinter()
+	perNodePodRuntime, err := o.PerNodePodOptions.ToRuntime(args)
 	if err != nil {
 		return nil, err
 	}
-	builder := o.ResourceBuilderFlags.ToBuilder(o.RESTClientGetter, args)
-	clientConfig, err := o.RESTClientGetter.ToRESTConfig()
-	if err != nil {
-		return nil, err
-	}
-	kubeClient, err := kubernetes.NewForConfig(clientConfig)
-	if err != nil {
-		return nil, err
-	}
-	imageClient, err := imagev1client.NewForConfig(clientConfig)
-	if err != nil {
-		return nil, err
-	}
-	imagePullSpec, err := resolveImageStreamTag(imageClient, "openshift", "must-gather", "latest")
-	if err != nil {
-		return nil, fmt.Errorf("unable to resolve image: %w", err)
-	}
-
-	ret := &RestartKubeletRuntime{
-		ResourceFinder: builder,
-		KubeClient:     kubeClient,
-
-		DryRun:                   o.DryRun,
-		ImagePullSpec:            imagePullSpec,
-		NumberOfNodesInParallel:  int(parallelInt),
-		PercentOfNodesInParallel: int(parallelPercentage),
-
-		Printer:   printer,
-		IOStreams: o.IOStreams,
-	}
-
-	return ret, nil
-}
-
-func resolveImageStreamTag(imageClient *imagev1client.ImageV1Client, namespace, name, tag string) (string, error) {
-	imageStream, err := imageClient.ImageStreams(namespace).Get(context.TODO(), name, metav1.GetOptions{})
-	if err != nil {
-		return "", err
-	}
-	var image string
-	if image, _, _, _, err = imageutil.ResolveRecentPullSpecForTag(imageStream, tag, false); err != nil {
-		return "", fmt.Errorf("unable to resolve the imagestream tag %s/%s:%s: %v", namespace, name, tag, err)
-	}
-	return image, nil
+	return &RestartKubeletRuntime{
+		PerNodePodRuntime: perNodePodRuntime,
+	}, nil
 }
